pyproject.toml
[tool.poetry]
name = "salesforce-observability"
version = "0.1.0"
description = ""
authors = ["Jônatas Lima de Medeiros <jonatas.lima.medeiros@live.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
prometheus-client = "^0.20.0"
simple-salesforce = "^1.12.6"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
